requirements.txt
numpy==1.24.2
pandas==1.5.3
nltk==3.8.1
scikit-learn==1.2.1
fastapi==0.93.0
uvicorn[standard]==0.21.0
mysql-connector-python==8.0.32
PyJWT==2.6.0
PyYAML==6.0
colorama==0.4.6